# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Finder v2",
    "description": "Finder v2 is a highly efficient image management system designed for organizing, storing, and preventing duplicate images in large databases.",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register new user",
        "description": "Create a new user account. On success, no body is returned.",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User registered successfully.",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "400": {
            "description": "Invalid payload or weak password policy violation."
          },
          "409": {
            "description": "Username or email already in use."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login and obtain tokens",
        "description": "Authenticate with username and password. Returns a short-lived access token and a long-lived refresh token.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Authentication successful. Tokens returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body."
          },
          "401": {
            "description": "Invalid username or password."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Refresh access token",
        "description": "Exchange a valid refresh token for a new access token. If the refresh token is invalid, expired, or revoked, an error is returned.",
        "operationId": "refresh_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "New access token issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body."
          },
          "401": {
            "description": "Invalid, expired, or revoked refresh token."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/collections/": {
      "post": {
        "tags": [
          "Collections"
        ],
        "summary": "Create collection",
        "description": "Create a new collection for the authenticated user.",
        "operationId": "create_collection_collections__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Collection created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionOut"
                }
              }
            }
          },
          "400": {
            "description": "Collection with this name already exists."
          },
          "401": {
            "description": "Unauthorized."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/collections/{collection_id}": {
      "patch": {
        "tags": [
          "Collections"
        ],
        "summary": "Update collection",
        "description": "Update name or tags of an existing collection.",
        "operationId": "update_collection_collections__collection_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Collection Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Collection updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionOut"
                }
              }
            }
          },
          "400": {
            "description": "Invalid update (for example, forbidden name)."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Collection not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Collections"
        ],
        "summary": "Delete collection",
        "description": "Delete a collection owned by the authenticated user.",
        "operationId": "delete_collection_collections__collection_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Collection Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Collection deleted."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Collection not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CollectionCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Name",
            "description": "Collection name.",
            "examples": [
              "Cats"
            ]
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "description": "Optional list of tags.",
            "examples": [
              [
                "sphynx",
                "2025",
                "tabby"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CollectionCreate",
        "description": "Payload to create a new collection."
      },
      "CollectionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Collection ID."
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id",
            "description": "Collection Owner ID."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Collection name."
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "Collection tags."
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "name"
        ],
        "title": "CollectionOut",
        "description": "Collection details."
      },
      "CollectionUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "New collection name.",
            "examples": [
              "Dogs"
            ]
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "description": "New list of tags.",
            "examples": [
              [
                "dachshund",
                "cute",
                "2013"
              ]
            ]
          }
        },
        "type": "object",
        "title": "CollectionUpdate",
        "description": "Payload to update an existing collection."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginIn": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Username chosen during registration.",
            "examples": [
              "diogo"
            ]
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "Plain-text password chosen during registration.",
            "examples": [
              "3w6mYtzE?D$d"
            ]
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginIn",
        "description": "Credentials used to obtain access and refresh tokens."
      },
      "RefreshIn": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token",
            "description": "Previously issued refresh token in JWT format.",
            "examples": [
              "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            ]
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshIn",
        "description": "Payload to rotate an existing (valid) refresh token."
      },
      "RegisterIn": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 32,
            "minLength": 3,
            "title": "Username",
            "description": "Unique username.",
            "examples": [
              "diogo"
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Valid and unique email address.",
            "examples": [
              "diogotoporcov@gmail.com"
            ]
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "description": "User password.",
            "examples": [
              "3w6mYtzE?D$d"
            ]
          }
        },
        "type": "object",
        "required": [
          "username",
          "email",
          "password"
        ],
        "title": "RegisterIn",
        "description": "Payload to create a new user account."
      },
      "TokenOut": {
        "properties": {
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "Token type. Always 'bearer'.",
            "default": "bearer",
            "examples": [
              "bearer"
            ]
          },
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "Short-lived JWT used to authorize API requests."
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "Access token lifetime in seconds.",
            "examples": [900]
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token",
            "description": "Long-lived JWT used to obtain new access tokens."
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in",
          "refresh_token"
        ],
        "title": "TokenOut",
        "description": "OAuth2-compatible bearer token response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}